Move image upload and predict logic to ImageContainer.
