+++ ClientApp/components/image-gallery.tsx
@@ -9,22 +9,21 @@ export interface ImageGalleryProps {
 
 export class ImageGallery extends React.Component<ImageGalleryProps> {
     public render(): JSX.Element | null {
-        if (this.props.images === null) {
-            return null;
-        }
         const rows: JSX.Element[] = [];
         for (let i = 0; i < this.props.images.length / 4; ++i) {
             const row: JSX.Element[] = [];
             for (let j = i * 4; j < i * 4 + 4 && j < this.props.images.length; ++j) {
                 row.push(
                     <ImageContainer
+                        key={`${this.props.images[j].name}-${i}`}
                         image={this.props.images[j]}
                         prediction={this.props.prediction}
+                        index={i}
                     />
                 );
             }
             rows.push(
-                <Row>
+                <Row key={i}>
                     {row}
                 </Row>
             );
